Long 타입 변수 -> Int 타입 변환, 최단시간 & 최소환승 알고리즘 구현

+++ src/main/java/com/busanit/busan_subway_project/controller/StationLocationController.java
@@ -13,7 +13,7 @@ public class StationLocationController {
     private StationLocationService stationLocationService;
 
     @GetMapping("stationlocation/{scode}")
-    public Station_Location getSLByScode(@PathVariable Long scode){
+    public Station_Location getSLByScode(@PathVariable Integer scode){
         return stationLocationService.getStationLocation(scode);
     }
